# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Prospix API
  description: |
    API multi-tenant da plataforma de prospecção inteligente Prospix.
    Primeiro tenant: Giovane Carrara (MetLife · SJRP).

    ## Autenticação
    Todas as rotas (exceto `/auth/*` e `/webhooks/*`) exigem JWT Bearer + header `X-Tenant-Id`.
    O `X-Tenant-Id` é validado contra o JWT — mismatch retorna 403.

    ## Multi-tenancy
    Todo dado é isolado por tenant via RLS no PostgreSQL. Um tenant nunca vê dados de outro.

    ## Rate limits
    - `/auth/*`: 10 req/min por IP
    - `/tenant/*` GET: 600 req/min por tenant
    - `/tenant/*` POST/PATCH: 60 req/min por tenant
  version: 1.0.0
  contact:
    name: Guilds
    email: gustavo.macedo@guilds.com.br
servers:
  - url: https://api.prospix.com.br/v1
    description: Produção
  - url: https://api-staging.prospix.com.br/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local dev
...
security:
  - bearerAuth: []
    tenantHeader: []
paths:
...
  /webhooks/evolution:
    post:
      tags: [Webhooks]
      summary: Dispatcher unificado da Evolution API
      x-prospix-response-shape: raw-object
      security: []
      parameters:
        - { name: x-evolution-signature, in: header, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: ACK rapido }
        '401': { description: HMAC invalido }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-Id
...